refactor(DataMapper): Split WrapperActionService
Fixes: https://github.com/KaotoIO/kaoto/issues/3559

 - Introduce `WrapperBaseService` for common routines
 - Migrate Abstract/Choice specific service methods to new `AbstractFieldService` and `ChoiceFieldService`
 - Migrate `computeAddFieldCandidates()` to a new `FieldCandidateService`
 - Promote `revertOverride()` into existing `FieldOverrideService`
 - Remove `WrapperActionService` as a result of the split
 - Restructure tests along with the refactoring
