+++ packages/ui/src/services/visualization/mapping-action.service.test.ts
@@ -15,13 +15,11 @@ import {
   ForEachItem,
   IfItem,
   MappingTree,
-  OtherwiseItem,
   UnknownMappingItem,
   ValueOfSelector,
   ValueOfType,
   ValueSelector,
   VariableItem,
-  WhenItem,
 } from '../../models/datamapper/mapping';
 import { MappingActionKind } from '../../models/datamapper/mapping-action';
 import { Types } from '../../models/datamapper/types';
@@ -31,7 +29,6 @@ import {
   FieldItemNodeData,
   FieldNodeData,
   MappingNodeData,
-  SourceVariableNodeData,
   TargetAbstractFieldNodeData,
   TargetChoiceFieldNodeData,
   TargetDocumentNodeData,
@@ -43,25 +40,16 @@ import {
 import { useDocumentTreeStore } from '../../store/document-tree.store';
 import {
   getConditionalMappingsToShipOrderXslt,
-  getContactsXsd,
-  getEnvelopeXsd,
-  getExtensionSimpleXsd,
-  getFieldSubstitutionXsd,
-  getOrderInfoXsd,
-  getOrgXsd,
-  getShipOrderToShipOrderInvalidForEachXslt,
   getShipOrderToShipOrderXslt,
   getUnknownApplyTemplateXslt,
   TestUtil,
 } from '../../stubs/datamapper/data-mapper';
 import { XmlSchemaCollection } from '../../xml-schema-ts';
 import { XmlSchemaDocument, XmlSchemaField } from '../document/xml-schema/xml-schema-document.model';
-import { XmlSchemaDocumentService } from '../document/xml-schema/xml-schema-document.service';
 import { MappingSerializerService } from '../mapping/mapping-serializer.service';
 import { MappingActionService } from './mapping-action.service';
 import { MappingActionRegistryService } from './mapping-action-registry.service';
 import { VisualizationService } from './visualization.service';
-import { WrapperActionService } from './wrapper-action.service';
 
 describe('MappingActionService', () => {
   let sourceDoc: XmlSchemaDocument;
@@ -82,760 +70,6 @@ describe('MappingActionService', () => {
   describe('without pre-populated mappings', () => {
     beforeEach(() => {
       targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-    });
-
-    describe('applyIf()', () => {
-      it('should add If', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        MappingActionService.applyIf(shipOrderChildren[0] as TargetNodeData);
-
-        expect(tree.children[0].name).toContain('fx-ShipOrder');
-        expect(tree.children[0].children[0].name).toBe('if');
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('if');
-        const ifChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        expect(ifChildren).toHaveLength(1);
-        expect(ifChildren[0].title).toBe('OrderId');
-      });
-
-      it('should add If on primitive target body', () => {
-        const primitiveTargetDoc = new PrimitiveDocument(
-          new DocumentDefinition(DocumentType.TARGET_BODY, DocumentDefinitionType.Primitive, BODY_DOCUMENT_ID),
-        );
-        tree = new MappingTree(
-          primitiveTargetDoc.documentType,
-          primitiveTargetDoc.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        targetDocNode = new TargetDocumentNodeData(primitiveTargetDoc, tree);
-        MappingActionService.applyIf(targetDocNode);
-
-        expect(VisualizationService.hasChildren(targetDocNode)).toBeTruthy();
-        let targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        expect(targetDocChildren).toHaveLength(1);
-        const ifItem = (targetDocChildren[0] as MappingNodeData).mapping;
-        expect(ifItem instanceof IfItem).toBeTruthy();
-        expect(ifItem.name).toBe('if');
-
-        targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        const ifChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(ifChildren).toHaveLength(1);
-        expect((ifChildren[0] as MappingNodeData).mapping instanceof ValueSelector).toBeTruthy();
-      });
-    });
-
-    describe('applyChooseWhenOtherwise()', () => {
-      it('should add Choose-When-Otherwise', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[1].title).toBe('OrderPerson');
-        MappingActionService.applyChooseWhenOtherwise(shipOrderChildren[1] as TargetNodeData);
-
-        expect(tree.children[0].name).toContain('fx-ShipOrder');
-        expect(tree.children[0].children[0].name).toBe('choose');
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-
-        expect(shipOrderChildren[1].title).toBe('choose');
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[1]);
-        expect(chooseChildren).toHaveLength(2);
-
-        expect(chooseChildren[0].title).toBe('when');
-        const whenChildren = VisualizationService.generateNonDocumentNodeDataChildren(chooseChildren[0]);
-        expect(whenChildren).toHaveLength(1);
-        const whenOrderPerson = whenChildren[0] as MappingNodeData;
-        expect(whenOrderPerson.title).toBe('OrderPerson');
-        expect(whenOrderPerson.mapping.parent instanceof WhenItem).toBeTruthy();
-
-        expect(chooseChildren[1].title).toBe('otherwise');
-        const otherwiseChildren = VisualizationService.generateNonDocumentNodeDataChildren(chooseChildren[1]);
-        expect(otherwiseChildren).toHaveLength(1);
-        const otherwiseOrderPerson = otherwiseChildren[0] as MappingNodeData;
-        expect(otherwiseOrderPerson.title).toBe('OrderPerson');
-        expect(otherwiseOrderPerson.mapping.parent instanceof OtherwiseItem).toBeTruthy();
-      });
-
-      it('should add Choose-When-Otherwise on primitive target body', () => {
-        const primitiveTargetDoc = new PrimitiveDocument(
-          new DocumentDefinition(DocumentType.TARGET_BODY, DocumentDefinitionType.Primitive, BODY_DOCUMENT_ID),
-        );
-        tree = new MappingTree(
-          primitiveTargetDoc.documentType,
-          primitiveTargetDoc.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        targetDocNode = new TargetDocumentNodeData(primitiveTargetDoc, tree);
-        MappingActionService.applyChooseWhenOtherwise(targetDocNode);
-
-        expect(VisualizationService.hasChildren(targetDocNode)).toBeTruthy();
-        let targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        expect(targetDocChildren).toHaveLength(1);
-        const chooseItem = (targetDocChildren[0] as MappingNodeData).mapping;
-        expect(chooseItem instanceof ChooseItem).toBeTruthy();
-        expect(chooseItem.name).toBe('choose');
-
-        targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(chooseChildren).toHaveLength(2);
-        const whenItem = (chooseChildren[0] as MappingNodeData).mapping;
-        expect(whenItem instanceof WhenItem).toBeTruthy();
-        expect(whenItem.children[0] instanceof ValueSelector).toBeTruthy();
-
-        const otherwiseItem = (chooseChildren[1] as MappingNodeData).mapping;
-        expect(otherwiseItem instanceof OtherwiseItem).toBeTruthy();
-        expect(otherwiseItem.children[0] instanceof ValueSelector).toBeTruthy();
-      });
-    });
-
-    describe('applyWhen()', () => {
-      it('should addWhen', () => {
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        MappingActionService.applyChooseWhenOtherwise(targetShipOrderChildren[1] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        MappingActionService.applyWhen(targetShipOrderChildren[1] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetShipOrderChildren[1]);
-        expect(chooseChildren).toHaveLength(3);
-
-        const whenItem1 = (chooseChildren[0] as MappingNodeData).mapping;
-        expect(whenItem1 instanceof WhenItem).toBeTruthy();
-        expect(whenItem1.children).toHaveLength(1);
-        expect(whenItem1.children[0] instanceof FieldItem).toBeTruthy();
-
-        const whenItem2 = (chooseChildren[1] as MappingNodeData).mapping;
-        expect(whenItem2 instanceof WhenItem).toBeTruthy();
-        expect(whenItem2.children).toHaveLength(1);
-        expect(whenItem2.children[0] instanceof FieldItem).toBeTruthy();
-
-        const otherwiseItem = (chooseChildren[2] as MappingNodeData).mapping;
-        expect(otherwiseItem instanceof OtherwiseItem).toBeTruthy();
-        expect(otherwiseItem.children).toHaveLength(1);
-        expect(otherwiseItem.children[0] instanceof FieldItem).toBeTruthy();
-      });
-
-      it('should add When in primitive target body choose', () => {
-        const primitiveTargetDoc = new PrimitiveDocument(
-          new DocumentDefinition(DocumentType.TARGET_BODY, DocumentDefinitionType.Primitive, BODY_DOCUMENT_ID),
-        );
-        tree = new MappingTree(
-          primitiveTargetDoc.documentType,
-          primitiveTargetDoc.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        targetDocNode = new TargetDocumentNodeData(primitiveTargetDoc, tree);
-        MappingActionService.applyChooseWhenOtherwise(targetDocNode);
-
-        let targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        MappingActionService.applyWhen(targetDocChildren[0] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(chooseChildren).toHaveLength(3);
-
-        const whenItem1 = (chooseChildren[0] as MappingNodeData).mapping;
-        expect(whenItem1 instanceof WhenItem).toBeTruthy();
-        expect(whenItem1.children).toHaveLength(1);
-        expect(whenItem1.children[0] instanceof ValueSelector).toBeTruthy();
-
-        const whenItem2 = (chooseChildren[1] as MappingNodeData).mapping;
-        expect(whenItem2 instanceof WhenItem).toBeTruthy();
-        expect(whenItem2.children).toHaveLength(1);
-        expect(whenItem2.children[0] instanceof ValueSelector).toBeTruthy();
-
-        const otherwiseItem = (chooseChildren[2] as MappingNodeData).mapping;
-        expect(otherwiseItem instanceof OtherwiseItem).toBeTruthy();
-        expect(otherwiseItem.children).toHaveLength(1);
-        expect(otherwiseItem.children[0] instanceof ValueSelector).toBeTruthy();
-      });
-    });
-
-    describe('applyOtherwise()', () => {
-      it('should add Otherwise', () => {
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        MappingActionService.applyChooseWhenOtherwise(targetShipOrderChildren[1] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        let chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetShipOrderChildren[1]);
-        MappingActionService.deleteMappingItem(chooseChildren[1] as MappingNodeData);
-        MappingActionService.applyOtherwise(targetShipOrderChildren[1] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetShipOrderChildren[1]);
-        expect(chooseChildren).toHaveLength(2);
-
-        const whenItem = (chooseChildren[0] as MappingNodeData).mapping;
-        expect(whenItem instanceof WhenItem).toBeTruthy();
-        expect(whenItem.children).toHaveLength(1);
-        expect(whenItem.children[0] instanceof FieldItem).toBeTruthy();
-
-        const otherwiseItem = (chooseChildren[1] as MappingNodeData).mapping;
-        expect(otherwiseItem instanceof OtherwiseItem).toBeTruthy();
-        expect(otherwiseItem.children).toHaveLength(1);
-        expect(otherwiseItem.children[0] instanceof FieldItem).toBeTruthy();
-      });
-
-      it('should add Otherwise in primitive target body choose', () => {
-        const primitiveTargetDoc = new PrimitiveDocument(
-          new DocumentDefinition(DocumentType.TARGET_BODY, DocumentDefinitionType.Primitive, BODY_DOCUMENT_ID),
-        );
-        tree = new MappingTree(
-          primitiveTargetDoc.documentType,
-          primitiveTargetDoc.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        targetDocNode = new TargetDocumentNodeData(primitiveTargetDoc, tree);
-        MappingActionService.applyChooseWhenOtherwise(targetDocNode);
-
-        let targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        let chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        MappingActionService.deleteMappingItem(chooseChildren[1] as MappingNodeData);
-        MappingActionService.applyOtherwise(targetDocChildren[0] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generatePrimitiveDocumentChildren(targetDocNode);
-        chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(chooseChildren).toHaveLength(2);
-
-        const whenItem = (chooseChildren[0] as MappingNodeData).mapping;
-        expect(whenItem instanceof WhenItem).toBeTruthy();
-        expect(whenItem.children).toHaveLength(1);
-        expect(whenItem.children[0] instanceof ValueSelector).toBeTruthy();
-
-        const otherwiseItem = (chooseChildren[1] as MappingNodeData).mapping;
-        expect(otherwiseItem instanceof OtherwiseItem).toBeTruthy();
-        expect(otherwiseItem.children).toHaveLength(1);
-        expect(otherwiseItem.children[0] instanceof ValueSelector).toBeTruthy();
-      });
-    });
-
-    describe('applyForEach()', () => {
-      it('should add for-each', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[3].title).toBe('Item');
-        MappingActionService.applyForEach(shipOrderChildren[3] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('for-each');
-        const forEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[3]);
-        expect(forEachChildren).toHaveLength(1);
-        expect(forEachChildren[0].title).toBe('Item');
-      });
-    });
-
-    describe('applyForEach() on abstract wrapper field', () => {
-      it('should wrap the abstract wrapper itself, not its parent', () => {
-        const NS_SUBSTITUTION = 'http://www.example.com/SUBSTITUTION';
-        const definition = new DocumentDefinition(
-          DocumentType.TARGET_BODY,
-          DocumentDefinitionType.XML_SCHEMA,
-          'test-doc',
-          { 'FieldSubstitution.xsd': getFieldSubstitutionXsd() },
-        );
-        definition.rootElementChoice = { namespaceUri: NS_SUBSTITUTION, name: 'Zoo' };
-        const result = XmlSchemaDocumentService.createXmlSchemaDocument(definition);
-        if (!result.document) throw new Error('Failed to create test document');
-        const document = result.document;
-
-        const mappingTree = new MappingTree(
-          document.documentType,
-          document.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        const docNode = new TargetDocumentNodeData(document, mappingTree);
-
-        const docChildren = VisualizationService.generateStructuredDocumentChildren(docNode);
-        const zooChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const equipmentNode = zooChildren.find((c) => c.title === 'Equipment');
-        expect(equipmentNode).toBeDefined();
-        const equipmentChildren = VisualizationService.generateNonDocumentNodeDataChildren(equipmentNode!);
-        const optionsNode = equipmentChildren.find((c) => c.title === 'Options');
-        expect(optionsNode).toBeDefined();
-        const optionsChildren = VisualizationService.generateNonDocumentNodeDataChildren(optionsNode!);
-        expect(optionsChildren).toHaveLength(1);
-        const abstractWrapperNode = optionsChildren[0];
-        expect(abstractWrapperNode).toBeInstanceOf(TargetAbstractFieldNodeData);
-
-        MappingActionService.applyForEach(abstractWrapperNode as TargetAbstractFieldNodeData);
-
-        expect(mappingTree.children).toHaveLength(1);
-        const zooFieldItem = mappingTree.children[0] as FieldItem;
-        expect(zooFieldItem).toBeInstanceOf(FieldItem);
-        expect(zooFieldItem.field.name).toBe('Zoo');
-
-        const equipmentFieldItem = zooFieldItem.children[0] as FieldItem;
-        expect(equipmentFieldItem.field.name).toBe('Equipment');
-        const optionsFieldItem = equipmentFieldItem.children[0] as FieldItem;
-        expect(optionsFieldItem.field.name).toBe('Options');
-
-        const optionsChild = optionsFieldItem.children[0];
-        expect(optionsChild).toBeInstanceOf(ForEachItem);
-        const forEachItem = optionsChild as ForEachItem;
-        expect(forEachItem.children).toHaveLength(1);
-        const wrappedFieldItem = forEachItem.children[0] as FieldItem;
-        expect(wrappedFieldItem).toBeInstanceOf(FieldItem);
-        expect(wrappedFieldItem.field.wrapperKind).toBe('abstract');
-      });
-    });
-
-    describe('applyForEachGroup()', () => {
-      it('should add for-each-group', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[3].title).toBe('Item');
-        MappingActionService.applyForEachGroup(shipOrderChildren[3] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('for-each-group');
-        const forEachGroupChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[3]);
-        expect(forEachGroupChildren).toHaveLength(1);
-        expect(forEachGroupChildren[0].title).toBe('Item');
-      });
-    });
-
-    describe('applyForEachCurrentGroup()', () => {
-      it('should wrap with for-each and set expression to current-group()', () => {
-        const docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('Item');
-
-        MappingActionService.applyForEachGroup(shipOrderChildren[3] as TargetFieldNodeData);
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedDocChildren[0]);
-        const forEachGroupNode = shipOrderChildren[3] as MappingNodeData;
-        expect(forEachGroupNode.title).toBe('for-each-group');
-
-        const forEachGroupChildren = VisualizationService.generateNonDocumentNodeDataChildren(forEachGroupNode);
-        const itemInsideGroup = forEachGroupChildren[0] as FieldItemNodeData;
-        expect(itemInsideGroup.title).toBe('Item');
-
-        MappingActionService.applyForEachCurrentGroup(itemInsideGroup);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const finalDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const finalShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(finalDocChildren[0]);
-        const finalForEachGroupNode = finalShipOrderChildren[3] as MappingNodeData;
-        const finalGroupChildren = VisualizationService.generateNonDocumentNodeDataChildren(finalForEachGroupNode);
-        const forEachNode = finalGroupChildren[0] as MappingNodeData;
-        expect(forEachNode.title).toBe('for-each');
-        expect((forEachNode.mapping as ForEachItem).expression).toBe('current-group()');
-      });
-    });
-
-    describe('applyInnerForEach()', () => {
-      it('should add inner for-each', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        MappingActionService.applyInnerForEach(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        // Should have at least one child which is the for-each
-        expect(orderIdChildren.length).toBeGreaterThanOrEqual(1);
-        const forEachChild = orderIdChildren.find((child) => child.title === 'for-each');
-        expect(forEachChild).toBeDefined();
-        expect(forEachChild?.title).toBe('for-each');
-      });
-
-      it('should add multiple inner for-each as siblings when applied to the same field', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-
-        // Add first inner for-each
-        MappingActionService.applyInnerForEach(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-
-        // Add second inner for-each to the same field (not to the for-each node)
-        MappingActionService.applyInnerForEach(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-
-        // Should have 2 for-each children as siblings
-        const forEachChildren = orderIdChildren.filter((child) => child.title === 'for-each');
-        expect(forEachChildren).toHaveLength(2);
-
-        // Both should be direct children of OrderId (siblings, not nested)
-        forEachChildren.forEach((forEachChild) => {
-          expect((forEachChild as MappingNodeData).mapping.parent).toBe(
-            (shipOrderChildren[0] as FieldItemNodeData).mapping,
-          );
-        });
-      });
-
-      it('should nest inner for-each when applied to an existing for-each node', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-
-        // Add first inner for-each
-        MappingActionService.applyInnerForEach(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const firstForEach = orderIdChildren.find((child) => child.title === 'for-each');
-
-        // Add second inner for-each to the for-each node itself (should nest)
-        MappingActionService.applyInnerForEach(firstForEach as TargetNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const updatedOrderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-
-        // Should still have only 1 for-each at the OrderId level
-        const topLevelForEach = updatedOrderIdChildren.filter((child) => child.title === 'for-each');
-        expect(topLevelForEach).toHaveLength(1);
-
-        // The nested for-each should be a child of the first for-each
-        const nestedChildren = VisualizationService.generateNonDocumentNodeDataChildren(topLevelForEach[0]);
-        const nestedForEach = nestedChildren.find((child) => child.title === 'for-each');
-        expect(nestedForEach).toBeDefined();
-        expect((nestedForEach as MappingNodeData).mapping.parent).toBe((topLevelForEach[0] as MappingNodeData).mapping);
-      });
-    });
-
-    describe('applyInnerForEachGroup()', () => {
-      it('should add inner for-each-group', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        MappingActionService.applyInnerForEachGroup(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        expect(orderIdChildren.length).toBeGreaterThanOrEqual(1);
-        const forEachGroupChild = orderIdChildren.find((child) => child.title === 'for-each-group');
-        expect(forEachGroupChild).toBeDefined();
-        expect(forEachGroupChild?.title).toBe('for-each-group');
-      });
-
-      it('should nest inner for-each-group when applied to an existing for-each node', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-
-        MappingActionService.applyInnerForEach(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const forEachNode = orderIdChildren.find((child) => child.title === 'for-each');
-
-        MappingActionService.applyInnerForEachGroup(forEachNode as TargetNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const updatedOrderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-
-        const topLevelForEach = updatedOrderIdChildren.filter((child) => child.title === 'for-each');
-        expect(topLevelForEach).toHaveLength(1);
-
-        const nestedChildren = VisualizationService.generateNonDocumentNodeDataChildren(topLevelForEach[0]);
-        const nestedForEachGroup = nestedChildren.find((child) => child.title === 'for-each-group');
-        expect(nestedForEachGroup).toBeDefined();
-        expect((nestedForEachGroup as MappingNodeData).mapping instanceof ForEachGroupItem).toBeTruthy();
-      });
-    });
-
-    describe('applyInnerForEachCurrentGroup()', () => {
-      it('should add inner for-each with expression current-group() inside for-each-group', () => {
-        const docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('Item');
-
-        MappingActionService.applyForEachGroup(shipOrderChildren[3] as TargetFieldNodeData);
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        let updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedDocChildren[0]);
-        const forEachGroupNode = shipOrderChildren[3] as MappingNodeData;
-        expect(forEachGroupNode.title).toBe('for-each-group');
-
-        MappingActionService.applyInnerForEachCurrentGroup(forEachGroupNode);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedDocChildren[0]);
-        const updatedGroupNode = shipOrderChildren[3] as MappingNodeData;
-        const groupChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedGroupNode);
-        const forEachNode = groupChildren.find((child) => child.title === 'for-each');
-        expect(forEachNode).toBeDefined();
-        expect((forEachNode as MappingNodeData).mapping instanceof ForEachItem).toBeTruthy();
-        expect(((forEachNode as MappingNodeData).mapping as ForEachItem).expression).toBe('current-group()');
-      });
-
-      it('should not be allowed outside for-each-group context', () => {
-        const docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('Item');
-
-        MappingActionService.applyForEach(shipOrderChildren[3] as TargetFieldNodeData);
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          updatedDocChildren[0],
-        );
-        const forEachNode = updatedShipOrderChildren[3] as MappingNodeData;
-        expect(forEachNode.title).toBe('for-each');
-        expect(MappingActionRegistryService.getAllowedActions(forEachNode)).not.toContain(
-          MappingActionKind.InnerForEachCurrentGroup,
-        );
-      });
-
-      it('should be allowed on for-each-group node', () => {
-        const docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('Item');
-
-        MappingActionService.applyForEachGroup(shipOrderChildren[3] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          updatedDocChildren[0],
-        );
-        const forEachGroupNode = updatedShipOrderChildren[3] as MappingNodeData;
-        expect(forEachGroupNode.title).toBe('for-each-group');
-        expect(MappingActionRegistryService.getAllowedActions(forEachGroupNode)).toContain(
-          MappingActionKind.InnerForEachCurrentGroup,
-        );
-      });
-
-      it('should not be allowed inside existing for-each current-group()', () => {
-        const docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[3].title).toBe('Item');
-
-        MappingActionService.applyForEachGroup(shipOrderChildren[3] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        let updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedDocChildren[0]);
-        const forEachGroupNode = updatedShipOrderChildren[3] as MappingNodeData;
-
-        MappingActionService.applyInnerForEachCurrentGroup(forEachGroupNode);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedDocChildren[0]);
-        const forEachGroupNode2 = updatedShipOrderChildren[3] as MappingNodeData;
-        const groupChildren = VisualizationService.generateNonDocumentNodeDataChildren(forEachGroupNode2);
-        const forEachCurrentGroupNode = groupChildren.find((child) => child.title === 'for-each') as MappingNodeData;
-        expect((forEachCurrentGroupNode.mapping as ForEachItem).expression).toBe('current-group()');
-
-        expect(MappingActionRegistryService.getAllowedActions(forEachCurrentGroupNode)).not.toContain(
-          MappingActionKind.InnerForEachCurrentGroup,
-        );
-      });
-    });
-
-    describe('applyInnerChooseWhenOtherwise()', () => {
-      it('should add inner choose-when-otherwise', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[1].title).toBe('OrderPerson');
-        MappingActionService.applyInnerChooseWhenOtherwise(shipOrderChildren[1] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[1].title).toBe('OrderPerson');
-        const orderPersonChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[1]);
-        // Should have at least one child which is the choose
-        expect(orderPersonChildren.length).toBeGreaterThanOrEqual(1);
-        const chooseChild = orderPersonChildren.find((child) => child.title === 'choose');
-        expect(chooseChild).toBeDefined();
-        expect(chooseChild?.title).toBe('choose');
-
-        // Verify choose has when and otherwise branches
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(chooseChild!);
-        expect(chooseChildren).toHaveLength(2);
-        expect(chooseChildren[0].title).toBe('when');
-        expect(chooseChildren[1].title).toBe('otherwise');
-      });
-
-      it('should add when with ValueSelector (not FieldItem) to inner choose', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[1].title).toBe('OrderPerson');
-        MappingActionService.applyInnerChooseWhenOtherwise(shipOrderChildren[1] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const orderPersonChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[1]);
-        const chooseChild = orderPersonChildren.find((child) => child.title === 'choose');
-
-        // Add a new when branch
-        MappingActionService.applyWhen(chooseChild as TargetNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const updatedOrderPersonChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          shipOrderChildren[1],
-        );
-        const updatedChooseChild = updatedOrderPersonChildren.find((child) => child.title === 'choose');
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedChooseChild!);
-
-        // Should have 3 children now: 2 when + 1 otherwise
-        expect(chooseChildren).toHaveLength(3);
-        expect(chooseChildren[0].title).toBe('when');
-        expect(chooseChildren[1].title).toBe('when');
-        expect(chooseChildren[2].title).toBe('otherwise');
-
-        // Verify the new when branch contains a ValueSelector, not a FieldItem
-        const newWhenChildren = VisualizationService.generateNonDocumentNodeDataChildren(chooseChildren[1]);
-        expect(newWhenChildren).toHaveLength(1);
-        expect((newWhenChildren[0] as MappingNodeData).mapping instanceof ValueSelector).toBeTruthy();
-      });
-
-      it('should add when with ValueSelector to inner choose nested inside inner for-each', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-
-        // First add inner for-each
-        MappingActionService.applyInnerForEach(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const forEachChild = orderIdChildren.find((child) => child.title === 'for-each');
-
-        // Then add inner choose inside the for-each
-        MappingActionService.applyInnerChooseWhenOtherwise(forEachChild as TargetNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const updatedOrderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const updatedForEachChild = updatedOrderIdChildren.find((child) => child.title === 'for-each');
-        const forEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedForEachChild!);
-        const chooseChild = forEachChildren.find((child) => child.title === 'choose');
-
-        // Add a new when branch to the choose that's inside the for-each
-        MappingActionService.applyWhen(chooseChild as TargetNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const finalOrderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const finalForEachChild = finalOrderIdChildren.find((child) => child.title === 'for-each');
-        const finalForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(finalForEachChild!);
-        const finalChooseChild = finalForEachChildren.find((child) => child.title === 'choose');
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(finalChooseChild!);
-
-        // Should have 3 children: 2 when + 1 otherwise
-        expect(chooseChildren).toHaveLength(3);
-        expect(chooseChildren[0].title).toBe('when');
-        expect(chooseChildren[1].title).toBe('when');
-        expect(chooseChildren[2].title).toBe('otherwise');
-
-        // Verify the new when branch contains a ValueSelector, not a FieldItem
-        const newWhenChildren = VisualizationService.generateNonDocumentNodeDataChildren(chooseChildren[1]);
-        expect(newWhenChildren).toHaveLength(1);
-        expect((newWhenChildren[0] as MappingNodeData).mapping instanceof ValueSelector).toBeTruthy();
-      });
-    });
-
-    describe('applyInnerIf()', () => {
-      it('should add inner if', () => {
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        expect(docChildren).toHaveLength(1);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren).toHaveLength(4);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        MappingActionService.applyInnerIf(shipOrderChildren[0] as TargetFieldNodeData);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        expect(shipOrderChildren[0].title).toBe('OrderId');
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        // Should have at least one child which is the if
-        expect(orderIdChildren.length).toBeGreaterThanOrEqual(1);
-        const ifChild = orderIdChildren.find((child) => child.title === 'if');
-        expect(ifChild).toBeDefined();
-        expect(ifChild?.title).toBe('if');
-      });
-
-      it('should add nested inner if inside an existing IfItem', () => {
-        // First add an inner if to OrderId
-        let docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        MappingActionService.applyInnerIf(shipOrderChildren[0] as TargetFieldNodeData);
-
-        // Get the created if node
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const orderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const ifChild = orderIdChildren.find((child) => child.title === 'if');
-        expect(ifChild).toBeDefined();
-
-        // Now apply inner if on the IfItem itself
-        MappingActionService.applyInnerIf(ifChild as TargetNodeData);
-
-        // Verify the nested structure
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        docChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(docChildren[0]);
-        const updatedOrderIdChildren = VisualizationService.generateNonDocumentNodeDataChildren(shipOrderChildren[0]);
-        const updatedIfChild = updatedOrderIdChildren.find((child) => child.title === 'if');
-
-        // The if should now have a nested if inside it
-        const nestedIfChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedIfChild!);
-        const nestedIf = nestedIfChildren.find((child) => child.title === 'if');
-        expect(nestedIf).toBeDefined();
-        expect(nestedIf?.title).toBe('if');
-      });
     });
 
     describe('applyValueSelector()', () => {
@@ -977,85 +211,6 @@ describe('MappingActionService', () => {
           updatedDocChildren[0],
         );
         expect(MappingActionService.hasValueOfSelector(updatedShipOrderChildren[0] as TargetNodeData)).toBe(false);
-      });
-    });
-
-    describe('removeParentContainerCopyOf via engageMapping', () => {
-      it('should remove copy-of when child field is mapped via DnD', () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const sourceItem = sourceShipOrderChildren[3] as FieldNodeData;
-
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-        MappingActionService.engageMapping(tree, sourceItem, targetItem);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const forEachNode = targetShipOrderChildren[3] as MappingNodeData;
-        const forEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(forEachNode);
-        const itemNode = forEachChildren[0] as FieldItemNodeData;
-        const itemMapping = itemNode.mapping;
-        expect(
-          itemMapping.children.some((c) => c instanceof CopyOfSelector && c.valueType === CopyOfType.CONTAINER),
-        ).toBe(true);
-
-        const sourceItemChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceShipOrderChildren[3]);
-        const sourceTitleField = sourceItemChildren[0] as FieldNodeData;
-        const targetItemChildren = VisualizationService.generateNonDocumentNodeDataChildren(itemNode);
-        const targetTitleField = targetItemChildren[0] as TargetFieldNodeData;
-        MappingActionService.engageMapping(tree, sourceTitleField, targetTitleField);
-
-        expect(
-          itemMapping.children.some((c) => c instanceof CopyOfSelector && c.valueType === CopyOfType.CONTAINER),
-        ).toBe(false);
-        expect(itemMapping.children.some((c) => c instanceof FieldItem && c.field.name === 'Title')).toBe(true);
-      });
-
-      it('should preserve explicit CONTAINER_NODE copy-of when child field is mapped via DnD', () => {
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetShipTo = targetShipOrderChildren.find(
-          (c) => c instanceof TargetFieldNodeData && c.field.name === 'ShipTo',
-        ) as TargetFieldNodeData;
-        MappingActionService.applyCopyOfSelector(targetShipTo);
-
-        const shipOrderMapping = tree.children[0] as FieldItem;
-        const shipToMapping = shipOrderMapping.children.find(
-          (c) => c instanceof FieldItem && c.field.name === 'ShipTo',
-        ) as FieldItem;
-        expect(
-          shipToMapping.children.some((c) => c instanceof CopyOfSelector && c.valueType === CopyOfType.CONTAINER_NODE),
-        ).toBe(true);
-
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const sourceShipTo = sourceShipOrderChildren.find(
-          (c) => c instanceof FieldNodeData && c.field.name === 'ShipTo',
-        ) as FieldNodeData;
-        const sourceShipToChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceShipTo);
-        const sourceNameField = sourceShipToChildren[0] as FieldNodeData;
-
-        const updatedTargetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(updatedTargetDocNode);
-        const updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          updatedDocChildren[0],
-        );
-        const updatedShipToNode = updatedShipOrderChildren.find(
-          (c) => c instanceof FieldItemNodeData && c.field.name === 'ShipTo',
-        ) as FieldItemNodeData;
-        const shipToNodeChildren = VisualizationService.generateNonDocumentNodeDataChildren(updatedShipToNode);
-        const targetNameField = shipToNodeChildren.find(
-          (c) => c instanceof TargetFieldNodeData && c.field.name === 'Name',
-        ) as TargetFieldNodeData;
-        MappingActionService.engageMapping(tree, sourceNameField, targetNameField);
-
-        expect(
-          shipToMapping.children.some((c) => c instanceof CopyOfSelector && c.valueType === CopyOfType.CONTAINER_NODE),
-        ).toBe(true);
-        expect(shipToMapping.children.some((c) => c instanceof FieldItem && c.field.name === 'Name')).toBe(true);
       });
     });
 
@@ -1394,509 +549,6 @@ describe('MappingActionService', () => {
         const expressionItem = VisualizationService.getExpressionItemForNode(targetDocNode);
         expect(expressionItem).toBeUndefined();
       });
-    });
-
-    describe('engageMapping()', () => {
-      it('should engage mapping to a MappingItem', () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        MappingActionService.applyIf(targetShipOrderChildren[1] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const ifItem = tree.children[0].children[0] as IfItem;
-        expect(ifItem.expression).toBe('');
-        MappingActionService.engageMapping(
-          tree,
-          sourceShipOrderChildren[1] as FieldNodeData,
-          targetShipOrderChildren[1] as TargetNodeData,
-        );
-
-        expect(ifItem.expression).toBe('/ns0:ShipOrder/ns0:OrderPerson');
-      });
-
-      it('should engage mapping to a Document', () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        expect(tree.children).toHaveLength(0);
-        MappingActionService.engageMapping(tree, sourceDocChildren[0] as FieldNodeData, targetDocNode);
-
-        expect(tree.children[0] instanceof ValueSelector).toBeTruthy();
-        const selector = tree.children[0] as ValueSelector;
-        expect(selector.expression).toBe('/ns0:ShipOrder');
-      });
-
-      it('should engage mapping to a field', () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(tree.children).toHaveLength(0);
-        MappingActionService.engageMapping(
-          tree,
-          sourceShipOrderChildren[1] as FieldNodeData,
-          targetShipOrderChildren[1] as TargetNodeData,
-        );
-
-        expect(tree.children[0] instanceof FieldItem).toBeTruthy();
-        expect(tree.children[0].children[0] instanceof FieldItem).toBeTruthy();
-        expect(tree.children[0].children[0].children[0] instanceof ValueSelector).toBeTruthy();
-        const selector = tree.children[0].children[0].children[0] as ValueSelector;
-        expect(selector.expression).toBe('/ns0:ShipOrder/ns0:OrderPerson');
-      });
-
-      it("should engage regular mapping even if it's dropped to a for-each wrapped collection field", () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const sourceItem = sourceShipOrderChildren[3] as FieldNodeData;
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-        MappingActionService.applyForEach(targetItem);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const forEach = targetShipOrderChildren[3] as MappingNodeData;
-        const forEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(forEach);
-        MappingActionService.engageMapping(tree, sourceItem, forEachChildren[0] as TargetFieldNodeData);
-
-        expect((forEach.mapping as ForEachItem).expression).toBe('');
-        expect(((forEachChildren[0] as FieldItemNodeData).mapping.children[0] as ValueSelector).expression).toBe(
-          '/ns0:ShipOrder/Item',
-        );
-      });
-
-      it('should not remove for-each targeted field item when selector is removed', () => {
-        MappingSerializerService.deserialize(getShipOrderToShipOrderInvalidForEachXslt(), targetDoc, tree, paramsMap);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        let forEachItem = (targetShipOrderChildren[3] as MappingNodeData).mapping as ForEachItem;
-        expect(forEachItem.children).toHaveLength(1);
-        expect(forEachItem.expression).toBe('');
-        let targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          targetShipOrderChildren[3],
-        );
-        let itemItem = targetForEachChildren[0] as TargetFieldNodeData;
-        expect((itemItem.mapping?.children[0] as ValueSelector).expression).toBe('/ns0:ShipOrder/Item');
-        MappingActionService.deleteMappingItem(targetForEachChildren[0] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        forEachItem = (targetShipOrderChildren[3] as MappingNodeData).mapping as ForEachItem;
-        expect(forEachItem).toBeDefined();
-        expect(forEachItem.children).toHaveLength(1);
-        expect(forEachItem.expression).toBe('');
-        targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetShipOrderChildren[3]);
-        itemItem = targetForEachChildren[0] as TargetFieldNodeData;
-        expect(itemItem.mapping).toBeDefined();
-        expect(itemItem.mapping?.children.length).toBe(0);
-      });
-
-      it('should not remove for-each targeted field item when descendent is removed', () => {
-        MappingSerializerService.deserialize(getShipOrderToShipOrderInvalidForEachXslt(), targetDoc, tree, paramsMap);
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        let targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          targetShipOrderChildren[3],
-        );
-        let targetItemChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetForEachChildren[0]);
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const sourceItemChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceShipOrderChildren[3]);
-        MappingActionService.deleteMappingItem(targetForEachChildren[0] as TargetNodeData);
-        MappingActionService.engageMapping(
-          tree,
-          sourceItemChildren[0] as FieldNodeData,
-          targetItemChildren[0] as TargetFieldNodeData,
-        );
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetShipOrderChildren[3]);
-        targetItemChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetForEachChildren[0]);
-        expect(((targetItemChildren[0] as TargetFieldNodeData).mapping?.children[0] as ValueSelector).expression).toBe(
-          '/ns0:ShipOrder/Item/Title',
-        );
-        MappingActionService.deleteMappingItem(targetItemChildren[0] as TargetNodeData);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const forEachItem = (targetShipOrderChildren[3] as MappingNodeData).mapping as ForEachItem;
-        expect(forEachItem).toBeDefined();
-        expect(forEachItem.children).toHaveLength(1);
-        expect(forEachItem.expression).toBe('');
-        targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetShipOrderChildren[3]);
-        targetItemChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetForEachChildren[0]);
-        expect((targetItemChildren[0] as TargetFieldNodeData).mapping?.children[0] as ValueSelector).toBeUndefined();
-      });
-
-      it('should generate correct XPath for fields with xs:extension', () => {
-        const adtInDocResult = TestUtil.createAdtInDoc();
-        const adtOutDocResult = TestUtil.createAdtOutDoc();
-        const adtTree = new MappingTree(
-          adtOutDocResult.document!.documentType,
-          adtOutDocResult.document!.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        const adtInDocNode = new DocumentNodeData(adtInDocResult.document!);
-        const adtOutDocNode = new TargetDocumentNodeData(adtOutDocResult.document!, adtTree);
-
-        const hl7Children = VisualizationService.generateStructuredDocumentChildren(adtInDocNode);
-        const hl7Node = hl7Children[0];
-        const hl7SubChildren = VisualizationService.generateNonDocumentNodeDataChildren(hl7Node) as FieldNodeData[];
-        const pidNode = hl7SubChildren.find((child) => child.field?.name === 'PID');
-        expect(pidNode).toBeDefined();
-
-        const pidChildren = VisualizationService.generateNonDocumentNodeDataChildren(pidNode!) as FieldNodeData[];
-        const field01Node = pidChildren.find((child) => child.field?.name === 'field_01');
-        expect(field01Node).toBeDefined();
-
-        const field01Children = VisualizationService.generateNonDocumentNodeDataChildren(
-          field01Node!,
-        ) as FieldNodeData[];
-        const comp02Node = field01Children.find((child) => child.field?.name === 'comp_02');
-        expect(comp02Node).toBeDefined();
-
-        MappingActionService.engageMapping(adtTree, comp02Node as FieldNodeData, adtOutDocNode);
-
-        const expressionItem = VisualizationService.getExpressionItemForNode(adtOutDocNode);
-        expect(expressionItem?.expression).toBe('/HL7/PID/field_01/comp_02');
-      });
-
-      it('should generate correct XPath for attributes with xs:extension', () => {
-        const extensionSimpleDef = new DocumentDefinition(
-          DocumentType.SOURCE_BODY,
-          DocumentDefinitionType.XML_SCHEMA,
-          BODY_DOCUMENT_ID,
-          { 'ExtensionSimple.xsd': getExtensionSimpleXsd() },
-        );
-        const sourceDocResult = XmlSchemaDocumentService.createXmlSchemaDocument(extensionSimpleDef);
-        const targetDocResult = XmlSchemaDocumentService.createXmlSchemaDocument(extensionSimpleDef);
-        const tree = new MappingTree(
-          targetDocResult.document!.documentType,
-          targetDocResult.document!.documentId,
-          DocumentDefinitionType.XML_SCHEMA,
-        );
-        const sourceDocNode = new DocumentNodeData(sourceDocResult.document!);
-        const targetDocNode = new TargetDocumentNodeData(targetDocResult.document!, tree);
-
-        const productChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const productNode = productChildren[0];
-        const productSubChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          productNode,
-        ) as FieldNodeData[];
-        const priceNode = productSubChildren.find((child) => child.field?.name === 'price');
-        expect(priceNode).toBeDefined();
-
-        const priceChildren = VisualizationService.generateNonDocumentNodeDataChildren(priceNode!) as FieldNodeData[];
-        const currencyAttr = priceChildren.find(
-          (child) => child.field?.name === 'currency' && child.field?.isAttribute,
-        );
-        expect(currencyAttr).toBeDefined();
-
-        MappingActionService.engageMapping(tree, currencyAttr as FieldNodeData, targetDocNode);
-
-        const expressionItem = VisualizationService.getExpressionItemForNode(targetDocNode);
-        expect(expressionItem?.expression).toBe('/ns0:Product/ns0:price/@currency');
-      });
-    });
-
-    describe('container auto-mapping', () => {
-      it('should place ForEachItem with copy-of in parent when both collections share name and namespace', () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const sourceItem = sourceShipOrderChildren[3] as FieldNodeData;
-
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-
-        MappingActionService.engageMapping(tree, sourceItem, targetItem);
-
-        const shipOrderFieldItem = tree.children[0] as FieldItem;
-        expect(shipOrderFieldItem).toBeInstanceOf(FieldItem);
-
-        const forEachItem = shipOrderFieldItem.children.find((c) => c instanceof ForEachItem) as ForEachItem;
-        expect(forEachItem).toBeDefined();
-        expect(forEachItem.expression).toBe('/ns0:ShipOrder/Item');
-
-        const innerFieldItem = forEachItem.children.find((c) => c instanceof FieldItem) as FieldItem;
-        expect(innerFieldItem).toBeDefined();
-        expect(innerFieldItem.field.name).toBe('Item');
-
-        const copyOf = innerFieldItem.children.find((c) => c instanceof ValueSelector) as ValueSelector;
-        expect(copyOf).toBeDefined();
-        expect(copyOf.expression).toBe('.');
-      });
-
-      it('should place ForEachItem with inner FieldItem when collections have different names', () => {
-        const orderInfoDef = new DocumentDefinition(
-          DocumentType.SOURCE_BODY,
-          DocumentDefinitionType.XML_SCHEMA,
-          BODY_DOCUMENT_ID,
-          { 'OrderInfo.xsd': getOrderInfoXsd() },
-        );
-        const orderInfoResult = XmlSchemaDocumentService.createXmlSchemaDocument(orderInfoDef);
-        expect(orderInfoResult.validationStatus).toBe('success');
-        const orderInfoDoc = orderInfoResult.document!;
-        const orderInfoDocNode = new DocumentNodeData(orderInfoDoc);
-
-        const orderInfoDocChildren = VisualizationService.generateStructuredDocumentChildren(orderInfoDocNode);
-        const orderInfoChildren = VisualizationService.generateNonDocumentNodeDataChildren(orderInfoDocChildren[0]);
-        const sourceOrderEntry = orderInfoChildren[0] as FieldNodeData;
-
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-
-        MappingActionService.engageMapping(tree, sourceOrderEntry, targetItem);
-
-        const shipOrderFieldItem = tree.children[0] as FieldItem;
-        expect(shipOrderFieldItem).toBeInstanceOf(FieldItem);
-
-        const forEachItem = shipOrderFieldItem.children.find((c) => c instanceof ForEachItem) as ForEachItem;
-        expect(forEachItem).toBeDefined();
-        expect(forEachItem.expression).toContain('OrderEntry');
-
-        const innerFieldItem = forEachItem.children.find((c) => c instanceof FieldItem) as FieldItem;
-        expect(innerFieldItem).toBeDefined();
-        expect(innerFieldItem.field.name).toBe('Item');
-
-        expect(innerFieldItem.children.length).toBeGreaterThan(0);
-        const titleMapping = innerFieldItem.children.find(
-          (c) => c instanceof FieldItem && c.field.name === 'Title',
-        ) as FieldItem;
-        expect(titleMapping).toBeDefined();
-      });
-
-      it('should not create duplicate ForEachItem on repeated DnD (copy-of case)', () => {
-        const sourceDocChildren = VisualizationService.generateStructuredDocumentChildren(sourceDocNode);
-        const sourceShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(sourceDocChildren[0]);
-        const sourceItem = sourceShipOrderChildren[3] as FieldNodeData;
-
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-
-        MappingActionService.engageMapping(tree, sourceItem, targetItem);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-
-        MappingActionService.engageMapping(tree, sourceItem, targetShipOrderChildren[3] as TargetNodeData);
-
-        const shipOrderFieldItem = tree.children[0] as FieldItem;
-        const forEachItems = shipOrderFieldItem.children.filter((c) => c instanceof ForEachItem);
-        expect(forEachItems).toHaveLength(1);
-      });
-
-      it('should not create duplicate ForEachItem on repeated DnD (auto-child case)', () => {
-        const orderInfoDef = new DocumentDefinition(
-          DocumentType.SOURCE_BODY,
-          DocumentDefinitionType.XML_SCHEMA,
-          BODY_DOCUMENT_ID,
-          { 'OrderInfo.xsd': getOrderInfoXsd() },
-        );
-        const orderInfoResult = XmlSchemaDocumentService.createXmlSchemaDocument(orderInfoDef);
-        const orderInfoDoc = orderInfoResult.document!;
-        const orderInfoDocNode = new DocumentNodeData(orderInfoDoc);
-
-        const orderInfoDocChildren = VisualizationService.generateStructuredDocumentChildren(orderInfoDocNode);
-        const orderInfoChildren = VisualizationService.generateNonDocumentNodeDataChildren(orderInfoDocChildren[0]);
-        const sourceOrderEntry = orderInfoChildren[0] as FieldNodeData;
-
-        let targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        let targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-
-        MappingActionService.engageMapping(tree, sourceOrderEntry, targetItem);
-
-        targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-
-        MappingActionService.engageMapping(tree, sourceOrderEntry, targetShipOrderChildren[3] as TargetNodeData);
-
-        const shipOrderFieldItem = tree.children[0] as FieldItem;
-        const forEachItems = shipOrderFieldItem.children.filter((c) => c instanceof ForEachItem);
-        expect(forEachItems).toHaveLength(1);
-      });
-    });
-
-    describe('engageMapping() with SourceVariableNodeData', () => {
-      it('should create VALUE mapping for select-form variable to document', () => {
-        const variable = new VariableItem(tree, 'myVar');
-        tree.children.push(variable);
-        const sourceNode = new SourceVariableNodeData(variable);
-
-        MappingActionService.engageMapping(tree, sourceNode, targetDocNode);
-
-        const vs = tree.children.find((c) => c instanceof ValueSelector) as ValueSelector;
-        expect(vs).toBeDefined();
-        expect(vs.expression).toBe('$myVar');
-      });
-
-      it('should create mapping for content-form variable to document', () => {
-        const variable = new VariableItem(tree, 'contentVar');
-        variable.children.push(new FieldItem(variable, targetDoc.fields[0]));
-        tree.children.push(variable);
-        const sourceNode = new SourceVariableNodeData(variable);
-
-        MappingActionService.engageMapping(tree, sourceNode, targetDocNode);
-
-        const vs = tree.children.find((c) => c instanceof ValueSelector) as ValueSelector;
-        expect(vs).toBeDefined();
-        expect(vs.expression).toBe('$contentVar');
-      });
-
-      it('should create mapping for variable to a target field', () => {
-        const sourceNode = new SourceVariableNodeData(new VariableItem(tree, 'myVar'));
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-
-        MappingActionService.engageMapping(tree, sourceNode, targetShipOrderChildren[1] as TargetNodeData);
-
-        const rootFieldItem = tree.children[0] as FieldItem;
-        expect(rootFieldItem).toBeInstanceOf(FieldItem);
-        const vs = rootFieldItem.children[0].children[0] as ValueSelector;
-        expect(vs.expression).toBe('$myVar');
-      });
-    });
-
-    describe('xs:anyType container mapping', () => {
-      it('should create copy-of with CONTAINER_NODE when xs:anyType source is mapped to a container target', () => {
-        const anyTypeDef = new DocumentDefinition(
-          DocumentType.SOURCE_BODY,
-          DocumentDefinitionType.XML_SCHEMA,
-          BODY_DOCUMENT_ID,
-          { 'Envelope.xsd': getEnvelopeXsd() },
-        );
-        const anyTypeResult = XmlSchemaDocumentService.createXmlSchemaDocument(anyTypeDef);
-        expect(anyTypeResult.validationStatus).toBe('success');
-        const anyTypeDoc = anyTypeResult.document!;
-        const anyTypeDocNode = new DocumentNodeData(anyTypeDoc);
-
-        const anyTypeDocChildren = VisualizationService.generateStructuredDocumentChildren(anyTypeDocNode);
-        const envelopeChildren = VisualizationService.generateNonDocumentNodeDataChildren(anyTypeDocChildren[0]);
-        const payloadField = envelopeChildren[0] as FieldNodeData;
-        expect(payloadField.title).toBe('Payload');
-
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const targetShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        const targetItem = targetShipOrderChildren[3] as TargetFieldNodeData;
-        expect(targetItem.title).toBe('Item');
-
-        MappingActionService.engageMapping(tree, payloadField, targetItem);
-
-        const shipOrderFieldItem = tree.children[0] as FieldItem;
-        expect(shipOrderFieldItem).toBeInstanceOf(FieldItem);
-
-        const itemFieldItem = shipOrderFieldItem.children.find(
-          (c) => c instanceof FieldItem && c.field.name === 'Item',
-        ) as FieldItem;
-        expect(itemFieldItem).toBeDefined();
-
-        const copyOf = itemFieldItem.children.find((c) => c instanceof CopyOfSelector) as CopyOfSelector;
-        expect(copyOf).toBeDefined();
-        expect(copyOf.valueType).toEqual(CopyOfType.CONTAINER_NODE);
-        expect(copyOf.expression).toMatch(/\/node\(\)$/);
-      });
-    });
-
-    it('should fill ContextItemExpr (.) and AbbrevReverseStep (..) in xpath when it maps under for-each', () => {
-      const orgDefinition = new DocumentDefinition(DocumentType.SOURCE_BODY, DocumentDefinitionType.XML_SCHEMA, 'Org', {
-        'Org.xsd': getOrgXsd(),
-      });
-      const orgResult = XmlSchemaDocumentService.createXmlSchemaDocument(orgDefinition);
-      expect(orgResult.validationStatus).toBe('success');
-      const orgDoc = orgResult.document!;
-      const contactsDefinition = new DocumentDefinition(
-        DocumentType.TARGET_BODY,
-        DocumentDefinitionType.XML_SCHEMA,
-        BODY_DOCUMENT_ID,
-        { 'Contacts.xsd': getContactsXsd() },
-      );
-      const contactsResult = XmlSchemaDocumentService.createXmlSchemaDocument(contactsDefinition);
-      expect(contactsResult.validationStatus).toBe('success');
-      const contactsDoc = contactsResult.document!;
-
-      const orgToContactsTree = new MappingTree(
-        contactsDoc.documentType,
-        contactsDoc.documentId,
-        DocumentDefinitionType.XML_SCHEMA,
-      );
-      const orgSourceNode = new DocumentNodeData(orgDoc);
-      const targetContactsNode = new TargetDocumentNodeData(contactsDoc, orgToContactsTree);
-
-      const orgSourceChildren = VisualizationService.generateStructuredDocumentChildren(orgSourceNode);
-      const orgChildren = VisualizationService.generateNonDocumentNodeDataChildren(orgSourceChildren[0]);
-      const orgNameField = orgChildren.find((f) => f.title === 'Name') as FieldNodeData;
-      const personChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-        orgChildren.find((f) => f.title === 'Person') as FieldNodeData,
-      );
-      const personNameField = personChildren.find((f) => f.title === 'Name') as FieldNodeData;
-      const emailField = personChildren.find((f) => f.title === 'Email') as FieldNodeData; // Email field
-
-      let targetContactsChildren = VisualizationService.generateStructuredDocumentChildren(targetContactsNode);
-      let contactsChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetContactsChildren[0]);
-
-      const targetContactField = contactsChildren[0] as TargetFieldNodeData; // Contact field
-      MappingActionService.applyForEach(targetContactField);
-
-      targetContactsChildren = VisualizationService.generateStructuredDocumentChildren(targetContactsNode);
-      contactsChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetContactsChildren[0]);
-      MappingActionService.engageMapping(orgToContactsTree, emailField, contactsChildren[0] as TargetNodeData);
-
-      targetContactsChildren = VisualizationService.generateStructuredDocumentChildren(targetContactsNode);
-      contactsChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetContactsChildren[0]);
-      let targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(contactsChildren[0]);
-      let targetContactChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetForEachChildren[0]);
-
-      let targetOrgNameField = targetContactChildren.find(
-        (child) => (child as TargetFieldNodeData).field?.name === 'OrgName',
-      ) as TargetFieldNodeData;
-      MappingActionService.engageMapping(orgToContactsTree, orgNameField, targetOrgNameField);
-      let targetPersonNameField = targetContactChildren.find(
-        (child) => (child as TargetFieldNodeData).field?.name === 'PersonName',
-      ) as TargetFieldNodeData;
-      MappingActionService.engageMapping(orgToContactsTree, personNameField, targetPersonNameField);
-      let targetEmailField = targetContactChildren.find(
-        (child) => (child as TargetFieldNodeData).field?.name === 'Email',
-      ) as TargetFieldNodeData;
-      MappingActionService.engageMapping(orgToContactsTree, emailField, targetEmailField);
-
-      targetContactsChildren = VisualizationService.generateStructuredDocumentChildren(targetContactsNode);
-      contactsChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetContactsChildren[0]);
-      targetForEachChildren = VisualizationService.generateNonDocumentNodeDataChildren(contactsChildren[0]);
-      targetContactChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetForEachChildren[0]);
-
-      targetOrgNameField = targetContactChildren.find(
-        (child) => (child as TargetFieldNodeData).field?.name === 'OrgName',
-      ) as TargetFieldNodeData;
-      expect(targetOrgNameField.mapping).toBeDefined();
-      let valueSelector = targetOrgNameField.mapping?.children[0] as ValueSelector;
-      expect(valueSelector).toBeDefined();
-      expect(valueSelector.expression).toBe('../../Name');
-
-      targetPersonNameField = targetContactChildren.find(
-        (child) => (child as TargetFieldNodeData).field?.name === 'PersonName',
-      ) as TargetFieldNodeData;
-      expect(targetPersonNameField.mapping).toBeDefined();
-      valueSelector = targetPersonNameField.mapping?.children[0] as ValueSelector;
-      expect(valueSelector).toBeDefined();
-      expect(valueSelector.expression).toBe('../Name');
-
-      targetEmailField = targetContactChildren.find(
-        (child) => (child as TargetFieldNodeData).field?.name === 'Email',
-      ) as TargetFieldNodeData;
-      expect(targetEmailField.mapping).toBeDefined();
-      valueSelector = targetEmailField.mapping?.children[0] as ValueSelector;
-      expect(valueSelector).toBeDefined();
-      expect(valueSelector.expression).toBe('.');
     });
   });
 
@@ -2282,72 +934,6 @@ describe('MappingActionService', () => {
         expect(shipOrderMappingItem.children[4] instanceof FieldItem).toBeTruthy();
         const itemItem = shipOrderMappingItem.children[4] as FieldItem;
         expect(itemItem.field.name).toBe('Item');
-      });
-    });
-
-    describe('applyIf() on AddMappingNodeData', () => {
-      it('should wrap with if when applied to an AddMappingNodeData', () => {
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(shipOrderChildren).toHaveLength(5);
-
-        expect(shipOrderChildren[4] instanceof AddMappingNodeData).toBeTruthy();
-        const addMappingNode = shipOrderChildren[4] as AddMappingNodeData;
-        expect(addMappingNode.title).toBe('Item');
-
-        const shipOrderMappingItem = targetDocNode.mappingTree.children[0];
-        const childCountBefore = shipOrderMappingItem.children.length;
-
-        MappingActionService.applyIf(addMappingNode);
-
-        expect(shipOrderMappingItem.children).toHaveLength(childCountBefore + 1);
-        const newChild = shipOrderMappingItem.children[childCountBefore];
-        expect(newChild instanceof IfItem).toBeTruthy();
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          updatedDocChildren[0],
-        );
-        const ifNodes = updatedShipOrderChildren.filter((c) => c.title === 'if');
-        const ifNode = ifNodes[ifNodes.length - 1];
-        expect(ifNode).toBeDefined();
-        const ifChildren = VisualizationService.generateNonDocumentNodeDataChildren(ifNode);
-        expect(ifChildren).toHaveLength(1);
-        expect(ifChildren[0].title).toBe('Item');
-      });
-    });
-
-    describe('applyChooseWhenOtherwise() on AddMappingNodeData', () => {
-      it('should wrap with choose-when-otherwise when applied to an AddMappingNodeData', () => {
-        const targetDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const shipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(targetDocChildren[0]);
-        expect(shipOrderChildren).toHaveLength(5);
-
-        expect(shipOrderChildren[4] instanceof AddMappingNodeData).toBeTruthy();
-        const addMappingNode = shipOrderChildren[4] as AddMappingNodeData;
-        expect(addMappingNode.title).toBe('Item');
-
-        const shipOrderMappingItem = targetDocNode.mappingTree.children[0];
-        const childCountBefore = shipOrderMappingItem.children.length;
-
-        MappingActionService.applyChooseWhenOtherwise(addMappingNode);
-
-        expect(shipOrderMappingItem.children).toHaveLength(childCountBefore + 1);
-        const newChild = shipOrderMappingItem.children[childCountBefore];
-        expect(newChild instanceof ChooseItem).toBeTruthy();
-
-        targetDocNode = new TargetDocumentNodeData(targetDoc, tree);
-        const updatedDocChildren = VisualizationService.generateStructuredDocumentChildren(targetDocNode);
-        const updatedShipOrderChildren = VisualizationService.generateNonDocumentNodeDataChildren(
-          updatedDocChildren[0],
-        );
-        const chooseNode = updatedShipOrderChildren.find((c) => c.title === 'choose');
-        expect(chooseNode).toBeDefined();
-        const chooseChildren = VisualizationService.generateNonDocumentNodeDataChildren(chooseNode!);
-        expect(chooseChildren).toHaveLength(2);
-        expect(chooseChildren[0].title).toBe('when');
-        expect(chooseChildren[1].title).toBe('otherwise');
       });
     });
 
@@ -2862,108 +1448,6 @@ describe('MappingActionService', () => {
           item.apply(freshAddMappingNode, { onUpdate, openModal });
         }).not.toThrow();
       }
-    });
-  });
-
-  describe('dispatchChoiceSelection / clearChoiceSelectionOnField — target-side FieldItem mutations', () => {
-    function createChoiceTargetSetup(maxOccurs: number | 'unbounded' = 'unbounded') {
-      const document = TestUtil.createTargetOrderDoc();
-      const mappingTree = new MappingTree(
-        document.documentType,
-        document.documentId,
-        DocumentDefinitionType.XML_SCHEMA,
-      );
-      const docNode = new TargetDocumentNodeData(document, mappingTree);
-      const parentField = document.fields[0];
-
-      const choiceField = new XmlSchemaField(parentField, 'testChoice', false);
-      choiceField.type = Types.Container;
-      choiceField.wrapperKind = 'choice';
-      choiceField.maxOccurs = maxOccurs;
-
-      const memberA = new XmlSchemaField(choiceField, 'memberA', false);
-      memberA.type = Types.String;
-      const memberB = new XmlSchemaField(choiceField, 'memberB', false);
-      memberB.type = Types.String;
-      choiceField.fields = [memberA, memberB];
-      parentField.fields.push(choiceField);
-
-      const fieldNode = new TargetFieldNodeData(docNode, choiceField);
-
-      return { document, mappingTree, docNode, choiceField, memberA, memberB, fieldNode };
-    }
-
-    describe('dispatchChoiceSelection — per-instance apply', () => {
-      it('should replace existing FieldItem field when mapping already exists', () => {
-        const { mappingTree, fieldNode, choiceField, memberB } = createChoiceTargetSetup();
-        const existingItem = new FieldItem(mappingTree, choiceField);
-        existingItem.isUserCreated = true;
-        mappingTree.children.push(existingItem);
-        fieldNode.mapping = existingItem;
-
-        WrapperActionService.dispatchChoiceSelection(fieldNode, choiceField, { memberIndex: 1 }, {}, true);
-
-        expect(mappingTree.children).toHaveLength(1);
-        const replaced = mappingTree.children[0] as FieldItem;
-        expect(replaced.field).toBe(memberB);
-        expect(replaced.isUserCreated).toBe(true);
-      });
-
-      it('should re-parent children when replacing existing FieldItem', () => {
-        const { mappingTree, fieldNode, choiceField, memberA, memberB } = createChoiceTargetSetup();
-        const existingItem = new FieldItem(mappingTree, choiceField);
-        existingItem.isUserCreated = true;
-        const childItem = new FieldItem(existingItem, memberA);
-        existingItem.children.push(childItem);
-        mappingTree.children.push(existingItem);
-        fieldNode.mapping = existingItem;
-
-        WrapperActionService.dispatchChoiceSelection(fieldNode, choiceField, { memberIndex: 1 }, {}, true);
-
-        const replaced = mappingTree.children[0] as FieldItem;
-        expect(replaced.field).toBe(memberB);
-        expect(replaced.isUserCreated).toBe(true);
-        expect(replaced.children).toHaveLength(1);
-        expect(replaced.children[0].parent).toBe(replaced);
-      });
-    });
-
-    describe('clearChoiceSelectionOnField — document-level clear', () => {
-      it('should remove FieldItem from parent on existing mapping', () => {
-        const { mappingTree, fieldNode, choiceField, memberA } = createChoiceTargetSetup(1);
-        const existingItem = new FieldItem(mappingTree, memberA);
-        existingItem.isUserCreated = true;
-        const childItem = new FieldItem(existingItem, memberA);
-        existingItem.children.push(childItem);
-        mappingTree.children.push(existingItem);
-        fieldNode.mapping = existingItem;
-
-        WrapperActionService.clearChoiceSelectionOnField(fieldNode, choiceField, {}, true);
-
-        expect(mappingTree.children).toHaveLength(0);
-        expect(existingItem.children).toHaveLength(0);
-      });
-
-      it('should revert FieldItem to wrapper field when inside InstructionItem', () => {
-        const { mappingTree, fieldNode, choiceField, memberA } = createChoiceTargetSetup(1);
-        const ifItem = new IfItem(mappingTree);
-        ifItem.expression = 'some-condition';
-        mappingTree.children.push(ifItem);
-
-        const existingItem = new FieldItem(ifItem, memberA);
-        existingItem.isUserCreated = true;
-        ifItem.children.push(existingItem);
-        fieldNode.mapping = existingItem;
-
-        WrapperActionService.clearChoiceSelectionOnField(fieldNode, choiceField, {}, true);
-
-        expect(ifItem.children).toHaveLength(1);
-        const reverted = ifItem.children[0] as FieldItem;
-        expect(reverted).toBeInstanceOf(FieldItem);
-        expect(reverted.field).toBe(choiceField);
-        expect(reverted.isUserCreated).toBe(true);
-        expect(ifItem.expression).toBe('some-condition');
-      });
     });
   });
 });
